pyproject.toml
[tool.poetry]
name = "screendetector"
version = "0.1.0"
description = ""
authors = ["NgankBakaa <ngankbakaa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
flask = "==3.1.1"
python-dotenv = "==1.1.0"
torch = "^2.7.0"
torchvision = "^0.22.0"
torchaudio = "^2.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
